3 new features
in the final snipe timings table I'd like a column titled 'Send In' that contains a timer showing you how long until you have to send (just creates a timer by taking the timespan between the send time and the current server time)

in the village selection add a field that contains a drop down with all of your groups. by default it should be set to 'all', but if the user selects another group (defense for example) it should filter the village selection rows to only villas in that group

next to target coordinates text box have a 'Your villas' (you can change the name) dropdown that lets the user select one of their villas). This should NOT be open by default. The user should have to click it for the dropdown to open